Compile and build once again
